# docs/swagger.yaml
basePath: /
definitions:
  common.ErrorResponse:
    properties:
      details:
        additionalProperties: true
        type: object
      error_code:
        type: string
      message:
        type: string
    type: object
...
  db.DeletionResponse:
    properties:
      id:
        type: string
    type: object
  db.NewQuickLaunch:
    properties:
      app_id:
        type: string
      app_version_id:
        type: string
      description:
        type: string
      is_public:
        type: boolean
      name:
        type: string
      submission:
        items:
          type: integer
        type: array
    type: object
...
  db.QuickLaunch:
    properties:
      app_id:
        type: string
      app_version_id:
        type: string
      creator:
        type: string
      description:
        type: string
      id:
        type: string
      is_public:
        type: boolean
      name:
        type: string
      submission:
        items:
          type: integer
        type: array
    type: object
...
  db.UpdateQuickLaunchRequest:
    properties:
      app_id:
        type: string
      app_version_id:
        type: string
      creator:
        type: string
      description:
        type: string
      is_public:
        type: boolean
      name:
        type: string
      submission:
        items:
          type: integer
        type: array
    type: object
...
host: localhost:60000
info:
  contact: {}
  description: API for the Discovery Environment's analyses, including Quick Launches
    and settings.
  license:
    name: 3-Clause BSD License
    url: https://github.com/cyverse-de/analyses?tab=License-1-ov-file#readme
  title: Analyses API
  version: "1.0"
paths:
...
  /quicklaunches:
    get:
      description: Gets all of the Quick Launches for a user. Includes UUIDs.
      parameters:
      - description: Username
        in: query
        name: user
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.QuickLaunch'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Get all Quick Launches
      tags:
      - quicklaunches
    post:
      consumes:
      - application/json
      description: Adds a Quick Launch and corresponding submission information to
        the database. The username passed in should already exist. A new UUID will
        be assigned and returned.
      parameters:
      - description: Username
        in: query
        name: user
        required: true
        type: string
      - description: Quick Launch to add
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/db.NewQuickLaunch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.QuickLaunch'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Add a Quick Launch
      tags:
      - quicklaunches
  /quicklaunches/{id}:
    delete:
      description: Deletes a Quick Launch from the database. Will return a success
        even if called on a Quick Launch that has either already been deleted or never
        existed.
      parameters:
      - description: Quick Launch ID
        in: path
        name: id
        required: true
        type: string
      - description: Username
        in: query
        name: user
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.DeletionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Delete a Quick Launch
      tags:
      - quicklaunches
    get:
      description: Gets the Quick Launch information from the database, including
        its UUID, the name of the user that owns it, and the submission JSON.
      parameters:
      - description: Quick Launch ID
        in: path
        name: id
        required: true
        type: string
      - description: Username
        in: query
        name: user
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.QuickLaunch'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Get a Quick Launch
      tags:
      - quicklaunches
    patch:
      consumes:
      - application/json
      description: Modifies an existing Quick Launch, allowing the caller to change
        owners and the contents of the submission JSON.
      parameters:
      - description: Quick Launch ID
        in: path
        name: id
        required: true
        type: string
      - description: Username
        in: query
        name: user
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/db.UpdateQuickLaunchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.QuickLaunch'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Update a Quick Launch
      tags:
      - quicklaunches
  /quicklaunches/{id}/app-info:
    get:
      description: Returns the app info needed to create and populate the app launcher
        in the UI. Populates the parameters with the values from the submission stored
        for the quick launch.
      parameters:
      - description: Quick Launch ID
        in: path
        name: id
        required: true
        type: string
      - description: Username
        in: query
        name: user
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Get Quick Launch app info
      tags:
      - quicklaunches
  /quicklaunches/apps/{id}:
    get:
      description: Returns a list of Quick Launches that the user can access based
        on the app's UUID.
      parameters:
      - description: App ID
        in: path
        name: id
        required: true
        type: string
      - description: Username
        in: query
        name: user
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.QuickLaunch'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Get Quick Launches by App
      tags:
      - quicklaunches
...
swagger: "2.0"